applications: nrf5340_audio: Refactor pscm
Moved the PCM stream channel modifier library
