+++ applications/nrf5340_audio/Kconfig.defaults
@@ -79,6 +79,10 @@ config TONE
 	bool
 	default y
 
+config PSCM
+	bool
+	default y
+
 # Enable NRFX_CLOCK for ACLK control
 config NRFX_CLOCK
 	bool
