+++ applications/nrf5340_audio/src/utils/CMakeLists.txt
@@ -9,6 +9,5 @@ target_sources(app PRIVATE
 	       ${CMAKE_CURRENT_SOURCE_DIR}/channel_assignment.c
 	       ${CMAKE_CURRENT_SOURCE_DIR}/data_fifo.c
 	       ${CMAKE_CURRENT_SOURCE_DIR}/error_handler.c
-	       ${CMAKE_CURRENT_SOURCE_DIR}/pcm_stream_channel_modifier.c
 	       ${CMAKE_CURRENT_SOURCE_DIR}/uicr.c
 )
