+++ applications/nrf5340_audio/src/utils/Kconfig
@@ -60,9 +60,5 @@ module = FW_INFO
 module-str = fw-info
 source "subsys/logging/Kconfig.template.log_config"
 
-module = PSCM
-module-str = pscm
-source "subsys/logging/Kconfig.template.log_config"
-
 endmenu # Log levels
 endmenu # Utils
