+++ applications/nrf5340_audio/src/utils/channel_assignment.h
@@ -14,6 +14,8 @@
  *
  */
 
+#include <audio_defines.h>
+
 #ifndef AUDIO_CHANNEL_DEFAULT
 #define AUDIO_CHANNEL_DEFAULT AUDIO_CH_L
 #endif /* AUDIO_CHANNEL_DEFAULT */
@@ -21,13 +23,6 @@
 static const char CH_L_TAG[] = "HL";
 static const char CH_R_TAG[] = "HR";
 static const char GW_TAG[] = "GW";
-
-/**@brief Audio channel assignment values */
-enum audio_channel {
-	AUDIO_CH_L,
-	AUDIO_CH_R,
-	AUDIO_CH_NUM,
-};
 
 /**
  * @brief Get assigned audio channel.
