+++ doc/nrf/releases/release-notes-changelog.rst
@@ -544,6 +544,11 @@ Other libraries
   * Updated by separating the library from the :ref:`nrf53_audio_app` application and moving it to :file:`lib/contin_array`.
     Updated code and documentation accordingly.
 
+* :ref:`lib_pcm_stream_channel_modifier` library:
+
+  * Separated the library from the :ref:`nrf53_audio_app` and moved it to :file:`lib/pcm_stream_channel_modifier`.
+    Updated code and documentation accordingly.
+
 * :ref:`QoS` library:
 
   * Updated by removing the ``QOS_MESSAGE_TYPES_REGISTER`` macro.
