+++ include/audio_defines.h
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2023 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+ */
+
+/** @file
+ *  @brief Globally accessible audio related defines
+ */
+
+#ifndef _AUDIO_DEFINES_H_
+#define _AUDIO_DEFINES_H_
+
+/**
+ * @brief Audio channel assignment values
+ */
+enum audio_channel {
+	AUDIO_CH_L,
+	AUDIO_CH_R,
+	AUDIO_CH_NUM,
+};
+
+#endif /* _AUDIO_DEFINES_H_ */
