+++ lib/CMakeLists.txt
@@ -43,3 +43,4 @@ add_subdirectory_ifdef(CONFIG_SFLOAT sfloat)
 add_subdirectory_ifdef(CONFIG_CONTIN_ARRAY contin_array)
 add_subdirectory_ifdef(CONFIG_PCM_MIX pcm_mix)
 add_subdirectory_ifdef(CONFIG_TONE tone)
+add_subdirectory_ifdef(CONFIG_PSCM pcm_stream_channel_modifier)
