+++ lib/Kconfig
@@ -44,5 +44,6 @@ rsource "sfloat/Kconfig"
 rsource "contin_array/Kconfig"
 rsource "pcm_mix/Kconfig"
 rsource "tone/Kconfig"
+rsource "pcm_stream_channel_modifier/Kconfig"
 
 endmenu
