+++ lib/pcm_stream_channel_modifier/CMakeLists.txt
@@ -1,0 +1,10 @@
+#
+# Copyright (c) 2023 Nordic Semiconductor
+#
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+#
+
+zephyr_library()
+zephyr_library_sources(
+	pcm_stream_channel_modifier.c
+)
