+++ tests/lib/pcm_stream_channel_modifier/prj.conf
@@ -1,2 +1,3 @@
 CONFIG_ZTEST=y
 CONFIG_IRQ_OFFLOAD=y
+CONFIG_PSCM=y
