--- tests/nrf5340_audio/pcm_stream_channel_modifier/CMakeLists.txt
@@ -1,22 +1,0 @@
-#
-# Copyright (c) 2018 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-cmake_minimum_required(VERSION 3.20.0)
-
-find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(NONE)
-
-target_sources(app
-  PRIVATE
-  main.c
-  ${ZEPHYR_NRF_MODULE_DIR}/applications/nrf5340_audio/src/utils/pcm_stream_channel_modifier.c
-  )
-
-target_include_directories(app
-  PRIVATE
-  ${ZEPHYR_NRF_MODULE_DIR}/applications/nrf5340_audio/src/utils/
-  ${ZEPHYR_NRF_MODULE_DIR}/applications/nrf5340_audio/src/audio/
-  )
