--- tests/nrf5340_audio/pcm_stream_channel_modifier/Kconfig
@@ -1,8 +1,0 @@
-# Copyright (c) 2022 Nordic Semiconductor ASA
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-
-module = PSCM
-module-str = pscm
-source "subsys/logging/Kconfig.template.log_config"
-
-source "Kconfig.zephyr"
